break up handler base class
